2021/02/02_test.go
package main

import (
	"log"
	"os"
	"testing"
)

func TestExamplesOneOne(t *testing.T) {
	tests := []struct {
		test     string
		expected int
	}{
		{`forward 5
down 5
forward 8
up 3
down 8
forward 2`, 150},
	}

	for _, test := range tests {
		result := run(test.test)
		if test.expected != result {
			t.Fatalf("Result % d != expected % d", result, test.expected)
		}
	}
}

func TestOneOne(t *testing.T) {
	file, err := os.ReadFile("./input")
	if err != nil {
		log.Fatalf("could not read file")
	}

	tests := []struct {
		test     string
		expected int
	}{
		{string(file), 2102357},
	}

	for _, test := range tests {
		result := run(test.test)
		if result != test.expected {
			t.Fatalf("Result % d != expected % d", result, test.expected)
		}
	}
}

func TestExamplesOneTwo(t *testing.T) {
	tests := []struct {
		test     string
		expected int
	}{
		{`forward 5
down 5
forward 8
up 3
down 8
forward 2`, 900},
		{`up 5
forward 10`, 500},
		{`down 10
forward 2`, 40},
	}

	for _, test := range tests {
		result := run_two(test.test)
		if test.expected != result {
			t.Fatalf("Result % d != expected % d", result, test.expected)
		}
	}
}

func TestOneTwo(t *testing.T) {
	file, err := os.ReadFile("./input")
	if err != nil {
		log.Fatalf("could not read file")
	}

	tests := []struct {
		test     string
		expected int
	}{
		{string(file), 2101031224},
	}

	for _, test := range tests {
		result := run_two(test.test)
		if result != test.expected {
			t.Fatalf("Result % d != expected % d", result, test.expected)
		}
	}
}
